requirements.txt
pyzmq==pyzmq-24.0.1